services/ai-worker/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
nats-py>=2.6.0
influxdb-client[async]>=1.40.0
scikit-learn>=1.3.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
websockets>=12.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
bcrypt==4.1.3
slowapi>=0.1.9
